Converts characters following the rules of 'neologd'

Converts characters following the rules of 'neologd'

str_jnormalize(str)

Arguments

str

Input vector.

Value

a character

Details

Converts the characters into normalized style basing on rules that is recommended by the Neologism dictionary for MeCab.

See also

Examples

str_jnormalize(
  paste0(
    "    \uff30",
    "\uff32\uff2d\uff2c\u300    \u526f    \u8aad    \u672c      "
  )
)
#> [1] "PRML̀副読本"
str_jnormalize(
  paste0(
    "\u5357\u30a2\u30eb\u30d7\u30b9\u306e\u3000\u5929\u7136\u6c34",
    "-\u3000\uff33\uff50\uff41\uff52\uff4b\uff49\uff4e\uff47\u3000",
    "\uff2c\uff45\uff4d\uff4f\uff4e\u3000\u30ec\u30e2\u30f3\u4e00\u7d5e\u308a"
 )
)
#> [1] "南アルプスの天然水-Sparking Lemonレモン一絞り"